requirements.txt

fastapi
uvicorn[standard]
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
docx2txt
pdfplumber
google-generativeai
pymupdf